[DSCP-502] move generic auth instances and functions from Educator.Web.Auth
