+++ core/package.yaml
@@ -56,7 +56,9 @@ library:
     - serialise
     - serokell-util
     - servant
+    - servant-auth-server
     - servant-server
+    - servant-client
     - servant-client-core
     - servant-generic
     - servant-util
@@ -71,6 +73,7 @@ library:
     - unliftio
     - unordered-containers
     - vinyl
+    - wai
     - yaml
 
 tests:
