+++ educator/src/Dscp/Educator/Web/Educator/Client/Logic.hs
@@ -13,11 +13,11 @@ import Servant.Generic (fromServant)
 import Servant.Util ()
 
 import Dscp.Core
-import Dscp.Educator.Web.Auth
 import Dscp.Educator.Web.Educator.API
 import Dscp.Educator.Web.Educator.Auth
 import Dscp.Educator.Web.Educator.Error
 import Dscp.Util
+import Dscp.Util.Servant.Auth
 import Dscp.Web
 
 -- | Exceptions which can appear from the client.
